requirements.txt
opencv-python
scipy
tensorflow==1.6.0
keras==2.1.5
h5py==2.10.0
ipython
pillow
matplotlib